test folder refactored
